+++ demos/cases/GanttDisplayMode.vue
@@ -2,7 +2,7 @@
 import { ref } from "vue";
 import { getData } from "../data";
 import { Gantt } from "../../src";
-import { Slider, RadioButtonGroup, Field, Button } from "@svar-ui/vue-core";
+import { Slider, RadioButtonGroup, Field } from "@svar-ui/vue-core";
 
 const props = defineProps({
 	skinSettings: {},
@@ -17,8 +17,6 @@ const options = [
 	{ id: "grid", label: "Grid" },
 	{ id: "chart", label: "Chart" },
 ];
-
-const api = ref(null);
 </script>
 
 <template>
@@ -34,22 +32,11 @@ const api = ref(null);
 			<Field :label="`Grid width: ${gridWidth}`" :position="'left'">
 				<Slider v-model:value="gridWidth" :min="200" :max="600" :step="50" />
 			</Field>
-			<div>
-				<Field label="For test only(action):" :position="'left'">
-					<Button
-						v-for="mode in ['all', 'grid', 'chart']"
-						:key="mode"
-						:onclick="() => api.exec('set-display-mode', { mode })"
-						>{{ mode }}</Button
-					>
-				</Field>
-			</div>
 		</div>
 
 		<div class="gtcell">
 			<Gantt
 				v-bind="skinSettings"
-				ref="api"
 				:tasks="data.tasks"
 				:links="data.links"
 				:scales="data.scales"