+++ demos/cases/GanttEditorCustomControls.vue
@@ -57,7 +57,7 @@ const data = getData();
 const api = ref(null);
 
 const columns = [
-	{ id: "text", header: "Task name", width: 220 },
+	{ id: "text", header: "Task name", flexgrow: 1 },
 	{ id: "assigned", header: "Assigned", width: 160, cell: AvatarCell },
 	{ id: "start", header: "Start Date", width: 100 },
 ];
@@ -72,6 +72,7 @@ const columns = [
 		:scales="data.scales"
 		lengthUnit="hour"
 		:columns="columns"
+		:gridWidth="480"
 	/>
 	<Editor :api="api" :items="items" />
 </template>