+++ demos/cases/GanttFilterInline.vue
@@ -34,7 +34,7 @@ const numberfilter = {
 };
 
 const columns = [
-	{ id: "text", header: ["Task name", textfilter], width: 200 },
+	{ id: "text", header: ["Task name", textfilter], flexgrow: 1 },
 	{
 		id: "start",
 		header: ["Start date", datefilter],
@@ -70,6 +70,7 @@ const columns = [
 					v-bind="skinSettings"
 					:tasks="tasks"
 					:columns="columns"
+					:gridWidth="600"
 					:links="data.links"
 					:scales="data.scales"
 					zoom