+++ src/components/chart/TimeScale.vue
@@ -103,7 +103,6 @@ const renderedRows = computed(() => {
 	right: 0px;
 	padding-left: 12px;
 	padding-right: 12px;
-	background-color: var(--wx-background);
 }
 
 .wx-cell {